Extract related helpers to internal trait
